Feat: Se agregaron las validaciones del buscar y eliminar junto con una imagen por default en nuestro ahorcado para las palabras nuevas que agregamos por medio de postman aparte se agregaron excepciones para los usos de puertos por jcruz-2021644

+++ ProyectoSpringAhorcado/src/main/java/com/jefrycruz/ProyectoSpringAhorcado/controller/PalabraController.java
@@ -21,8 +21,12 @@ public class PalabraController {
     }
 
     @GetMapping("/{codigoPalabra}")
-    public Palabra getPalabraById(@PathVariable Integer codigoPalabra){
-        return palabraService.getPalabraById(codigoPalabra);
+    public Object getPalabraById(@PathVariable Integer codigoPalabra){
+        Palabra palabra = palabraService.getPalabraById(codigoPalabra);
+        if (palabra == null){
+            return "No se encontro ninguna palabra con ese id";
+        }
+        return palabra;
     }
 
     @PostMapping
@@ -64,6 +68,10 @@ public class PalabraController {
 
     @DeleteMapping("/{codigoPalabra}")
     public String deletePalabra(@PathVariable Integer codigoPalabra){
+        Palabra palabra = palabraService.getPalabraById(codigoPalabra);
+        if (palabra == null){
+            return "No se encontro ninguna palabra con ese id para eliminar";
+        }
         palabraService.deletePalabra(codigoPalabra);
         return "Palabra y pista eliminada correctamente";
     }
